requirements.txt
flask==3.0.3
flask-cors==4.0.1
bcrypt==4.1.3
cryptography==42.0.8
pyotp==2.9.0
APScheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
